final touch

+++ src/Components/Brandcarinfoo.jsx
@@ -1,19 +1,41 @@
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2'
 
-const Brandcarinfoo = ({ datum }) => {
+const Brandcarinfoo = ({ datum, item, setitem }) => {
     const { _id, brand, imageurl, name, price } = datum;
 
+
     const handleDeleteItem = _id => {
-        fetch(`http://localhost:5000/cardetails/${brand}/${_id}`, {
-            method: "DELETE",
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+                fetch(`http://localhost:5000/cardetails/${brand}/${_id}`, {
+                    method: "DELETE",
+                })
+                    .then(res => {
+                        res.json();
+                    })
+                    .then(() => {
+                        const remaining = item.filter(itemm => itemm._id !== _id);
+                        setitem(remaining)
+                    })
+            }
         })
-            .then(res => {
-                res.json()
-            })
-            .catch(error => {
-                console.log(error);
-            })
     }
